gh-93649: Move PyFrame C API tests to test_capi
* Move C API tests from test_frame to test_capi.test_frame.
* Add Modules/_testcapi/frame.c file.
* Add Lib/test/test_capi/test_frame.py file.
