+++ Modules/_testcapimodule.c
@@ -2534,109 +2534,6 @@ test_tstate_capi(PyObject *self, PyObject *Py_UNUSED(args))
 }
 
 static PyObject *
-frame_getlocals(PyObject *self, PyObject *frame)
-{
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-    return PyFrame_GetLocals((PyFrameObject *)frame);
-}
-
-static PyObject *
-frame_getglobals(PyObject *self, PyObject *frame)
-{
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-    return PyFrame_GetGlobals((PyFrameObject *)frame);
-}
-
-static PyObject *
-frame_getgenerator(PyObject *self, PyObject *frame)
-{
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-    return PyFrame_GetGenerator((PyFrameObject *)frame);
-}
-
-static PyObject *
-frame_getbuiltins(PyObject *self, PyObject *frame)
-{
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-    return PyFrame_GetBuiltins((PyFrameObject *)frame);
-}
-
-static PyObject *
-frame_getlasti(PyObject *self, PyObject *frame)
-{
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-    int lasti = PyFrame_GetLasti((PyFrameObject *)frame);
-    if (lasti < 0) {
-        assert(lasti == -1);
-        Py_RETURN_NONE;
-    }
-    return PyLong_FromLong(lasti);
-}
-
-static PyObject *
-frame_new(PyObject *self, PyObject *args)
-{
-    PyObject *code, *globals, *locals;
-    if (!PyArg_ParseTuple(args, "OOO", &code, &globals, &locals)) {
-        return NULL;
-    }
-    if (!PyCode_Check(code)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a code object");
-        return NULL;
-    }
-    PyThreadState *tstate = PyThreadState_Get();
-
-    return (PyObject *)PyFrame_New(tstate, (PyCodeObject *)code, globals, locals);
-}
-
-static PyObject *
-test_frame_getvar(PyObject *self, PyObject *args)
-{
-    PyObject *frame, *name;
-    if (!PyArg_ParseTuple(args, "OO", &frame, &name)) {
-        return NULL;
-    }
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-
-    return PyFrame_GetVar((PyFrameObject *)frame, name);
-}
-
-static PyObject *
-test_frame_getvarstring(PyObject *self, PyObject *args)
-{
-    PyObject *frame;
-    const char *name;
-    if (!PyArg_ParseTuple(args, "Oy", &frame, &name)) {
-        return NULL;
-    }
-    if (!PyFrame_Check(frame)) {
-        PyErr_SetString(PyExc_TypeError, "argument must be a frame");
-        return NULL;
-    }
-
-    return PyFrame_GetVarString((PyFrameObject *)frame, name);
-}
-
-
-static PyObject *
 gen_get_code(PyObject *self, PyObject *gen)
 {
     if (!PyGen_Check(gen)) {
@@ -3599,14 +3496,6 @@ static PyMethodDef TestMethods[] = {
     {"type_get_tp_mro", type_get_tp_mro, METH_O},
     {"get_basic_static_type", get_basic_static_type, METH_VARARGS, NULL},
     {"test_tstate_capi", test_tstate_capi, METH_NOARGS, NULL},
-    {"frame_getlocals", frame_getlocals, METH_O, NULL},
-    {"frame_getglobals", frame_getglobals, METH_O, NULL},
-    {"frame_getgenerator", frame_getgenerator, METH_O, NULL},
-    {"frame_getbuiltins", frame_getbuiltins, METH_O, NULL},
-    {"frame_getlasti", frame_getlasti, METH_O, NULL},
-    {"frame_new", frame_new, METH_VARARGS, NULL},
-    {"frame_getvar", test_frame_getvar, METH_VARARGS, NULL},
-    {"frame_getvarstring", test_frame_getvarstring, METH_VARARGS, NULL},
     {"gen_get_code", gen_get_code, METH_O, NULL},
     {"get_feature_macros", get_feature_macros, METH_NOARGS, NULL},
     {"test_code_api", test_code_api, METH_NOARGS, NULL},
@@ -4402,6 +4291,9 @@ PyInit__testcapi(void)
         return NULL;
     }
     if (_PyTestCapi_Init_Import(m) < 0) {
+        return NULL;
+    }
+    if (_PyTestCapi_Init_Frame(m) < 0) {
         return NULL;
     }
