+++ PCbuild/_testcapi.vcxproj
@@ -128,6 +128,7 @@
     <ClCompile Include="..\Modules\_testcapi\monitoring.c" />
     <ClCompile Include="..\Modules\_testcapi\config.c" />
     <ClCompile Include="..\Modules\_testcapi\import.c" />
+    <ClCompile Include="..\Modules\_testcapi\frame.c" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc" />
