+++ PCbuild/_testcapi.vcxproj.filters
@@ -117,6 +117,9 @@
     <ClCompile Include="..\Modules\_testcapi\import.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Modules\_testcapi\frame.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc">
